% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPLFit.R
\name{FPLFit}
\alias{FPLFit}
\title{This function determines the 4 parameter log fit constants for each protein in the melt shift data sets.
See the Inflect function file for an example of program that could be executed to test the program function.}
\usage{
FPLFit(
  Data_Norm_Omit,
  NormBothCorrect,
  Condition,
  Temperature,
  NumberTemperatures
)
}
\arguments{
\item{Data_Norm_Omit}{List of accession numbers with NA omitted}

\item{NormBothCorrect}{List of normalized values from the FPLFit Correction function}

\item{Condition}{The Condition or the Control depending on which data set is being analyzed}

\item{Temperature}{The temperatures from the heat treatment}

\item{NumberTemperatures}{The number of temperatures in the heat treatment}
}
\value{
normalized data for each protein to the Inflect program
}
\description{
This function determines the 4 parameter log fit constants for each protein in the melt shift data sets.
See the Inflect function file for an example of program that could be executed to test the program function.
}
