% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPLFit_Correction.R
\name{FPLFit_Correction}
\alias{FPLFit_Correction}
\title{This function determines the 4 parameter log fit constants for the two conditions in a replicate}
\usage{
FPLFit_Correction(Median, Data_Norm_Omit, Condition, Temperature)
}
\arguments{
\item{Median}{The median fold change abundance from each temperature}

\item{Data_Norm_Omit}{List of accession numbers with NA omitted}

\item{Condition}{The Condition or the Control depending on which data set is being analyzed}

\item{Temperature}{The temperatures from the heat treatment}
}
\value{
normalized data to the Inflect program
}
\description{
This function determines the 4 parameter log fit constants for the two conditions in a replicate
}
